исправил замечания

+++ src/main/java/ru/yandex/practicum/filmorate/controllers/FilmController.java
@@ -47,8 +47,8 @@ public class FilmController {
     }
 
     @GetMapping("/common")
-    public List<Film> getCommonFilms(@RequestParam(value = "userId") Integer userId,
-                                     @RequestParam(value = "friendId") Integer friendId) {
+    public List<Film> getCommonFilms(@RequestParam(value = "userId") Long userId,
+                                     @RequestParam(value = "friendId") Long friendId) {
         return filmService.getCommonFilms(userId, friendId);
     }
